requirements.txt
asgiref==3.2.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
coverage==5.1
Django==3.0.1
django-cors-headers==3.2.0
djangorestframework==3.11.0
dnspython==1.16.0
eventlet==0.25.1
greenlet==0.4.15
gunicorn==20.0.4
httpretty==1.0.2
idna==2.8
lark-parser==0.8.6
monotonic==1.5
more-itertools==8.4.0
mysqlclient==1.4.6
packaging==20.4
Pillow==6.2.1
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.2
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
python-decouple==3.3
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
wcwidth==0.2.4
